+++ scripts/applications/gm-sidebar/gm-sidebar-base.mjs
@@ -962,11 +962,7 @@ export class GMSidebarAppBase extends foundry.applications.api.HandlebarsApplica
   }
 
   static _onLaunchSceneMode() {
-    const state = game.storyframe.stateManager?.getState();
-    if (!state?.speakers?.length) {
-      ui.notifications.warn(game.i18n.localize('STORYFRAME.CinematicScene.NoSpeakers'));
-      return;
-    }
+
     game.storyframe.socketManager.launchSceneMode();
   }